[TE-5476] Merge upload-commit-metadata into backfill-commit-metadata as --upload flag
Replace the separate upload-commit-metadata subcommand with a --upload
flag on backfill-commit-metadata, per review feedback. The two operations
are the same workflow (generate + upload) split across invocations, so a
single command with mode dispatch is a better UX.

When --upload is set, BackfillCommitMetadata dispatches to a private
uploadOnly helper that skips all git work and goes straight to scope
verification, presign, and S3 upload. Validation branches on UploadFile:
upload mode only requires API connection fields (no suite slug, days,
concurrency).

- Add --upload flag to backfill-commit-metadata (replaces --file on the
  removed upload-commit-metadata subcommand)
- Remove upload-commit-metadata subcommand, handler, validation method,
  flags, and dedicated command/test files
- Update ValidateForBackfillCommitMetadata to branch on UploadFile
- Add uploadOnly private function with the upload-only pipeline
- Move upload tests into backfill test file (3 tests: happy path, file
  not found, scope check failure)
- Update docs and README to reference --upload instead of the separate
  command
